Use static_cast; resolve more warnings
Mostly just fixing typecasting errors. The worst
offenders would be the handling of HWND and HMENU
types, as they change size depending on the
platform.

+++ Source/Action.cpp
@@ -102,12 +102,12 @@ Action *History::GetLastAction() const
 
 int History::GetUndoLevel() const
 {
-	return m_UndoStack.size();
+	return static_cast<int>(m_UndoStack.size());
 }
 
 int History::GetRedoLevel() const
 {
-	return m_RedoStack.size();
+	return static_cast<int>(m_RedoStack.size());
 }
 
 bool History::CanUndo() const
